refactor: AlertDialogのPropsを追加 #144
共通使用出来るようにするため
